refactor: @frontend, main 경로에서 user로 수정
