+++ apps/frontend/src/app/user/page.tsx
@@ -1,8 +1,8 @@
 import styles from './page.module.css';
 import Text from '@repo/ui/Text';
-import FormSection from '@/_components/pages/main/FormSection';
+import FormSection from '@/_components/pages/user/FormSection';
 
-export default function MainPage() {
+export default function Page() {
   return (
     <div className={styles.container}>
       <Text as="h2" size="lg" align="center" className={styles.title}>
